+++ test/react-native/features/fixtures/scenario-launcher/lib/CommandRunner.js
@@ -1,19 +1,33 @@
 import { getMazeRunnerAddress } from './ConfigFileReader'
 
-const RETRY_COUNT = 20
+const DEFAULT_RETRY_COUNT = 20
 const INTERVAL = 500
 
 let mazeAddress
 
 const delay = ms => new Promise(resolve => setTimeout(resolve, ms))
 
-const fetchCommand = async (url) => {
+export async function getCurrentCommand (allowedRetries = DEFAULT_RETRY_COUNT) {
+  if (allowedRetries <= 0) {
+    throw new Error(`allowedRetries must be a number >0, got '${allowedRetries}'`)
+  }
+
+  if (!mazeAddress) {
+    mazeAddress = await getMazeRunnerAddress()
+  }
+
+  const url = `http://${mazeAddress}/command`
+  console.error(`[BugsnagPerformance] Fetching command from ${url}`)
+
   let retries = 0
 
-  while (retries++ < RETRY_COUNT) {
+  while (retries++ < allowedRetries) {
     try {
       const response = await fetch(url)
-      const command = await response.json()
+      const text = await response.text()
+      console.error(`[BugsnagPerformance] Response from maze runner: ${text}`)
+
+      const command = JSON.parse(text)
 
       // keep polling until a scenario command is received
       if (command.action !== 'noop') {
@@ -25,24 +39,10 @@ const fetchCommand = async (url) => {
       console.error(`[BugsnagPerformance] Error fetching command from maze runner: ${err.message}`, err)
     }
 
-    console.error(`[BugsnagPerformance] ${RETRY_COUNT - retries} retries remaining...`)
+    console.error(`[BugsnagPerformance] ${allowedRetries - retries} retries remaining...`)
 
     await delay(INTERVAL)
   }
 
   throw new Error('Retry limit exceeded, giving up...')
 }
-
-const getCurrentCommand = async () => {
-  if (!mazeAddress) {
-    mazeAddress = await getMazeRunnerAddress()
-  }
-
-  const commandUrl = `http://${mazeAddress}/command`
-  console.error(`[BugsnagPerformance] Fetching command from ${commandUrl}`)
-
-  const command = await fetchCommand(commandUrl)
-  return command
-}
-
-module.exports.getCurrentCommand = getCurrentCommand